+++ src/blocks.js
@@ -158,7 +158,7 @@ CustomCommandBlockMorph, SymbolMorph, ToggleButtonMorph, DialMorph*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.blocks = '2020-December-04';
+modules.blocks = '2020-December-07';
 
 var SyntaxElementMorph;
 var BlockMorph;
@@ -2839,7 +2839,6 @@ BlockMorph.prototype.userMenu = function () {
     ) {
         alternatives = {
             reportMap : 'reportAtomicMap',
-			reportMap : 'reportAtomicMultimap',
             reportKeep : 'reportAtomicKeep',
             reportFindFirst: 'reportAtomicFindFirst',
             reportCombine : 'reportAtomicCombine'
@@ -2854,7 +2853,7 @@ BlockMorph.prototype.userMenu = function () {
         contains(
             [
                 'reportAtomicMap',
-				'reportAtomicMultimap',				
+                'reportAtomicMultimap',
                 'reportAtomicKeep',
                 'reportAtomicFindFirst',
                 'reportAtomicCombine'
@@ -2864,7 +2863,7 @@ BlockMorph.prototype.userMenu = function () {
     ) {
         alternatives = {
             reportAtomicMap : 'reportMap',
-			reportAtomicMultimap : 'reportMap',
+            reportAtomicMultimap : 'reportMap',
             reportAtomicKeep : 'reportKeep',
             reportAtomicFindFirst: 'reportFindFirst',
             reportAtomicCombine : 'reportCombine'
@@ -11717,7 +11716,7 @@ MultiArgMorph.prototype.is3ArgRingInHOF = function () {
                     [
                         'reportMap',
                         'reportAtomicMap',
-						'reportAtomicMultimap',
+                        'reportAtomicMultimap',
                         'reportKeep',
                         'reportAtomicKeep',
                         'reportFindFirst',